third_party/blink/web_tests/external/wpt/css/css-fonts/variations/font-shorthand-expected.txt

This is a testharness.js-based test.
[FAIL] Font shorthand: Font weight specified as calc(), value smaller than 1
  assert_equals: Font shorthand expected weight: Font weight specified as calc(), value smaller than 1 expected "1" but got "0.25"
[FAIL] Font shorthand: Font weight specified as calc(), value greater than 1000
  assert_equals: Font shorthand expected weight: Font weight specified as calc(), value greater than 1000 expected "1000" but got "1000.5"
[FAIL] Font shorthand: 'oblique' without slant angle
  assert_equals: Font shorthand expected style: 'oblique' without slant angle expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by valid small weight
  assert_equals: Font shorthand expected style: 'oblique' followed by valid small weight expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by valid large weight
  assert_equals: Font shorthand expected style: 'oblique' followed by valid large weight expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by valid calc() weight
  assert_equals: Font shorthand expected style: 'oblique' followed by valid calc() weight expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by a to-be-clamped calc() weight
  assert_equals: Font shorthand expected style: 'oblique' followed by a to-be-clamped calc() weight expected "oblique" but got "italic"
[FAIL] Font shorthand: calc() weight folowed by 'oblique'
  assert_equals: Font shorthand expected style: calc() weight folowed by 'oblique' expected "oblique" but got "italic"
[FAIL] Font shorthand: To-be-clamped calc() weight folowed by 'oblique' and slant angle
  assert_equals: Font shorthand expected weight: To-be-clamped calc() weight folowed by 'oblique' and slant angle expected "1000" but got "1200"
Harness: the test ran to completion.
